% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desconto.r
\name{desconto}
\alias{desconto}
\title{Calculo de Desconto.}
\usage{
desconto()
}
\description{
Siga as instrucoes da tela e forneca as informacoes solicitadas para obter o resultado.
}
\examples{
desconto()
}
\author{
Alexandre Barbosa | acbdev@hotmail.com | +55 9.8370-8251
}
