% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jurocomposto.r
\name{jurocomposto}
\alias{jurocomposto}
\title{Calcula juros compostos.}
\usage{
jurocomposto(c, i, t)
}
\arguments{
\item{c, }{Capital investido.}

\item{i, }{A taxa de juros representada por um numero inteiro.}

\item{t, }{quantidade de periodos, ex: quantidade de meses.}
}
\description{
Calcula juros compostos.
}
\examples{
jurocomposto(1000,5,12)
}
\author{
Alexandre Barbosa | acbdev@hotmail.com | +55 9.8370-8251
}
