% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juros.r
\name{juros}
\alias{juros}
\title{Calcula juros tanto simples como compostos.}
\usage{
juros()
}
\description{
Calcula juros tanto simples como compostos, basta seguir as instrucoes apresentadas na tela fornecendo as informacoes solicitadas.
}
\author{
Alexandre Barbosa | acbdev@hotmail.com | +55 31 9.8370-8251
}
