% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jurosimples.r
\name{jurosimples}
\alias{jurosimples}
\title{Calcula juros simples.}
\usage{
jurosimples(c, i, t)
}
\arguments{
\item{c, }{o valor do capital inicial.}

\item{i, }{A taxa de juros.}

\item{t, }{a quantidade de periodos envolvidos, ex: quantidade de meses.}
}
\description{
Faz o calculo de juros simples.
}
\examples{
jurosimples(1000,5,12)
}
\author{
Alexandre Barbosa | acbdev@hotmail.com | +55 31 9.8370-8251
}
