% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melhorpreco.r
\name{melhorpreco}
\alias{melhorpreco}
\title{Comparacao de preco entre 2 produtos}
\usage{
melhorpreco()
}
\description{
Informe o preco total e a quantidade de 2 produtos e tenha o retorno de qual e o mais economico.
}
\details{
Siga as instrucoes fornecidas na tela para obter o resultado.
}
\author{
Alexandre Barbosa | acbdev@hotmail.com | +55 31 9.8370-8251
}
