% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regtres.r
\name{regtres}
\alias{regtres}
\title{Calcula regra de 3 simples.}
\usage{
regtres(a, b, c)
}
\arguments{
\item{a, }{primeira grandeza}

\item{b, }{segunda grandeza}

\item{c, }{terceira grandeza}
}
\description{
Calcula regra de 3 simples.
}
\examples{
regtres(10, 20, 100)

}
\author{
Alexandre Barbosa | acbdev@hotmail.com | +55 9.8370-8251
}
