% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resto.r
\name{resto}
\alias{resto}
\title{Calcula o resto da divisao.}
\usage{
resto(x, y)
}
\arguments{
\item{x, }{dividendo.}

\item{y, }{divisor.}
}
\description{
Calcula o resto da divisao.
}
\examples{
resto(11,5)
}
\author{
Alexandre Barbosa | acbdev@hotmail.com | +55 9.8370-8251
}
