% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jurocomposto.r, R/regtres.r
\name{soma}
\alias{soma}
\alias{jurocomposto}
\alias{regtres}
\title{Faz a Soma de Dois Numeros}
\usage{
jurocomposto(c, itemp, t)

regtres(a, b, c)
}
\arguments{
\item{x}{Um numero}

\item{y}{Outro numero}
}
\value{
A soma dos numeros \code{x} e \code{y}.

A soma dos numeros \code{x} e \code{y}.
}
\description{
Uma (incrivel) funcao que pega dois numeros e faz a
soma. Utilize este campo para descrever melhor o proposito de
sua funcao e o que ela e capaz de fazer.

Uma (incrivel) funcao que pega dois numeros e faz a
soma. Utilize este campo para descrever melhor o proposito de
sua funcao e o que ela e capaz de fazer.
}
\details{
Utilize este campo para escrever detalhes mais tecnicos da
sua funcao (se necessario), ou para detalhar melhor como
utilizar determinados argumentos.

Utilize este campo para escrever detalhes mais tecnicos da
sua funcao (se necessario), ou para detalhar melhor como
utilizar determinados argumentos.
}
\examples{
soma(2, 2)

x <- 3
y <- 4
soma(x = x, y = y)

soma(2, 2)

x <- 3
y <- 4
soma(x = x, y = y)

}
\seealso{
\code{\link[base]{sum}}, \code{\link[base]{+}}

\code{\link[base]{sum}}, \code{\link[base]{+}}
}
\author{
Fernando Mayer

Fernando Mayer
}
