% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabprice.r
\name{tabprice}
\alias{tabprice}
\title{Calcula valores da tabela Price.}
\usage{
tabprice()
}
\description{
Siga as instrucoes da tela e forneca as informacoes solicitadas para obter o resultado.
}
\author{
Alexandre Barbosa | acbdev@hotmail.com | +55 9.8370-8251
}
