% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabsac.r
\name{tabsac}
\alias{tabsac}
\title{Calcula valores da tabela SAC.}
\usage{
tabsac()
}
\description{
Siga as instrucoes da tela e forneca as informacoes solicitadas para obter o resultado.
}
\author{
Alexandre Barbosa | acbdev@hotmail.com | +55 9.8370-8251
}
