% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabuada.r
\name{tabuada}
\alias{tabuada}
\title{Exibe a tabuada do número fornecido.}
\usage{
tabuada(x)
}
\arguments{
\item{x}{O número o qual deseja-se saber a tabuada.}
}
\description{
Exibe a tabuada do número fornecido exibindo os produtos da multiplicação dele iniciando pelo número 1 até o número 10.
}
\seealso{
\code{\link[base]{sum}}, \code{\link[base]{+}}
}
\author{
Alexandre Barbosa | acbdev@hotmail.com | +55 31 9.8370-8251
}
