% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apresentar.r
\name{apresentar}
\alias{apresentar}
\title{Calcula valores da tabela SAC.}
\usage{
apresentar()
}
\description{
Lista de funcoes contidas no pacote
desconto(); Calculo de descontos.
juros(); Calcula juros tanto simples como compostos, conforme escolha do usuario.
jurocomposto(); Calcula juros compostos.
jurosimples(); Calcula juros simples.
melhorpreco(); Determina qual produto entre 2 é o mais economico.
regtres(); Calcula regra de 3 simples.
resto(); Apresenta o resto da divisao exata, mas apresenta tambem o quociente da divisao com parte decimal.
tabprice(); Calcula o valor da prestacao pelo metodo de tabela Price.
tabsac(); Calcula o valor das prestações pelo metodo de tabela SAC apresentando uma tabela na tela.
tabuda(); Exibe a tabuada do numero solicitado.
Todas funões possuem ajuda disponível digitando o ponto de interrogacao antes de seus nomes sem os parenteses.
}
\author{
Alexandre Barbosa | acbdev@hotmail.com | +55 9.8370-8251
}
