% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabuada.r
\name{tabuada}
\alias{tabuada}
\title{Exibe a tabuada do numero fornecido.}
\usage{
tabuada(x)
}
\arguments{
\item{x}{O numero o qual deseja-se saber a tabuada.}
}
\description{
Exibe a tabuada do numero fornecido exibindo os produtos da multiplicacao dele, iniciando pelo numero 1 ate o numero 10.
}
\author{
Alexandre Barbosa | acbdev@hotmail.com | +55 31 9.8370-8251
}
